+++ Covid19Radar/Covid19Radar.Android/MainActivity.cs
@@ -34,6 +34,7 @@ namespace Covid19Radar.Droid
             base.OnCreate(savedInstanceState);
 
             Xamarin.Essentials.Platform.Init(this, savedInstanceState);
+            global::Rg.Plugins.Popup.Popup.Init(this, savedInstanceState);
             Xamarin.Forms.Forms.Init(this, savedInstanceState);
             Xamarin.Forms.FormsMaterial.Init(this, savedInstanceState);
