+++ Covid19Radar/Covid19Radar/ViewModels/HelpPage/HelpPage1ViewModel.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections.Generic;
+using Covid19Radar.Model;
+using Covid19Radar.Renderers;
+using Covid19Radar.Views;
+using Prism.Navigation;
+using Xamarin.Forms;
+
+namespace Covid19Radar.ViewModels
+{
+    public class HelpPage1ViewModel : ViewModelBase
+    {
+        public HelpPage1ViewModel(INavigationService navigationService) : base(navigationService)
+        {
+            Title = Resources.AppResources.TitleHowItWorks;
+        }
+
+    }
+}
