--- Covid19Radar/Covid19Radar/ViewModels/Tutorial/HelpPage4ViewModel.cs
@@ -1,22 +1,0 @@
-﻿using System.Collections.Generic;
-using Covid19Radar.Model;
-using Covid19Radar.Renderers;
-using Covid19Radar.Views;
-using Prism.Navigation;
-using Xamarin.Forms;
-
-namespace Covid19Radar.ViewModels
-{
-    public class HelpPage4ViewModel : ViewModelBase
-    {
-        public HelpPage4ViewModel(INavigationService navigationService) : base(navigationService)
-        {
-            Title = Resources.AppResources.TitleHowItWorks;
-        }
-
-        public Command OnClickNext => new Command(async () =>
-        {
-            await NavigationService.NavigateAsync(nameof(HelpPage5));
-        });
-    }
-}
