+++ Covid19Radar/Covid19Radar/Views/HelpPage/HelpMenuPage.xaml
@@ -1,9 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage
-    x:Class="Covid19Radar.Views.HelpPage1"
+<popup:PopupPage
+    x:Class="Covid19Radar.Views.HelpMenuPage"
     xmlns="http://xamarin.com/schemas/2014/forms"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:ios="clr-namespace:Xamarin.Forms.PlatformConfiguration.iOSSpecific;assembly=Xamarin.Forms.Core"
+    xmlns:popup="clr-namespace:Rg.Plugins.Popup.Pages;assembly=Rg.Plugins.Popup"
     xmlns:prism="http://prismlibrary.com"
     xmlns:template="clr-namespace:Covid19Radar.Templates"
     Title="使い方"
@@ -11,7 +12,6 @@
     prism:ViewModelLocator.AutowireViewModel="True"
     Style="{StaticResource DefaultPageStyle}"
     Visual="Material">
-
     <StackLayout
         Padding="10"
         BackgroundColor="#CCCCCC"
@@ -33,4 +33,4 @@
             </ListView.Behaviors>
         </ListView>
     </StackLayout>
-</ContentPage>
+</popup:PopupPage>
