+++ Covid19Radar/Covid19Radar/Views/HelpPage/HelpMenuPage.xaml.cs
@@ -1,12 +1,13 @@
-﻿using Xamarin.Forms;
+﻿using Rg.Plugins.Popup.Pages;
+using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 
 namespace Covid19Radar.Views
 {
     [XamlCompilation(XamlCompilationOptions.Compile)]
-    public partial class HelpPage5 : ContentPage
+    public partial class HelpMenuPage : PopupPage
     {
-        public HelpPage5()
+        public HelpMenuPage()
         {
             InitializeComponent();
         }
