+++ Covid19Radar/Covid19Radar/Views/HelpPage/HelpPage1.xaml
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage
-    x:Class="Covid19Radar.Views.HelpPage2"
+    x:Class="Covid19Radar.Views.HelpPage1"
     xmlns="http://xamarin.com/schemas/2014/forms"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:ffimageloading="clr-namespace:FFImageLoading.Forms;assembly=FFImageLoading.Forms"
