+++ Covid19Radar/Covid19Radar/Views/HelpPage/HelpPage1.xaml.cs
