+++ Covid19Radar/Covid19Radar/Views/HelpPage/HelpPage2.xaml.cs
